user == null bug should be fixed

+++ Maturarbeit android project/app/src/main/java/ch/imlee/maturarbeit/game/events/gameStateEvents/GameStartEvent.java
@@ -92,7 +92,7 @@ public class GameStartEvent extends GameStateEvent {
         return temp;
     }
 
-    public int getUserID(){
+    public byte getUserID(){
         return userID;
     }
